feat!(vad-trt): merge files to ROS package
