+++ AV-Solutions/vad-trt/app/autoware_tensorrt_vad/lib/vad_model.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "vad_model.hpp"
+#include "autoware/tensorrt_vad/vad_model.hpp"
 
 namespace autoware::tensorrt_vad {
