+++ AV-Solutions/vad-trt/app/include/autoware/tensorrt_vad/vad_interface_config.hpp
@@ -1,0 +1,1 @@
+../../../autoware_tensorrt_vad/include/autoware/tensorrt_vad/vad_interface_config.hpp
